fix(blink): create mapping to disable auto-inserting brackets on accept
I've been wanting this for a while. The completion auto-inserts brackets
which is usually fine _except_ when I'm passing a function as an
argument to another function. In that case I have to delete the brackets
and it's a bit annoying. So now I have a mapping <C-y> that will
accept the completion but not auto-insert brackets.